Remove unnecessary `IPTablesProxy` struct
Since we are allowed to implement our traits for foreign types, and Rust
is smart enough to correctly resolve function calls within a trait to
functions with the same signature in the actual type, we can get rid of
the `IPTablesProxy` struct.

Since `IPTablesProxy` was part of the public API of the dfw-library,
this is a breaking change and requires a major bump (or in our case:
from 0.4.x to 0.5.x).