Replace deprecated `chan`/`chan-signal`
This change replaces the deprecated packages `chan` and `chan-signal`
with `crossbeam-channel` and `signal_hook` respectively.

This is in accordance to the deprecation notice:
https://users.rust-lang.org/t/ann-chan-is-deprecated-use-crossbeam-channel-instead/19251.

It also changes how the spawned threads are handled: rather than
cleaning them through relatively complicated methods, we'll just do
nothing instead and let the OS correctly reap them once we exit the main
process.

This fixes issue #109.