Use `append_replace` on non-DFW chains
Instead of simply `append`ing to chains that are not fully managed by
DFW, we'll use `append_replace` to make sure we only append rules to
external chains once.

This fixes issue #97.

Although the `iptables-restore` backend delivers this feature "for
free", this change also implements for `IPTablesRestore::append_replace`
for completeness.