Refactor Docker integration tests
This change deduplicates the common test-resources used by the nftables
and iptables backends as much as possible.

It also modifies how tests can be added through the `dfw_tests!()`
macro. It uses the paste-crate to enable identifier concatenation and
thus simplifies the call-site.

This commit also prepares additional syntax for specifying both
regression tests and tests that should be failing.