Make exposing containers via IPv6 configurable
This change brings the iptables-backend to the same feature-set as the
nftables-backend: IPv6 is now supported through host-mounted ports,
rules allowing the traffic will be generated automatically.

This also introduces an `expose_via_ipv6`-flag for
wider-world-to-container rules to allow an end-user to disable the
behaviour of opening the IPv6 firewall.