Always populate `DFW::global_defaults`
The behaviour of DFW changes based on whether the section exists or not.
Given that none of the keys within the section are mandatory, just
adding `[global_defaults]` to the configuration will cause a change in
the rules generated (specifically: the
`default_docker_bridge_to_host_policy` field will be prepopulated with
an "accept" and thus a rule will be created for this).

Please note that adding this change means that there is a change in
DFW's default behaviour, meaning there is a default accept generated for
the main Docker bridge (if available), although most users of DFW should
not be affected by this since they probably provided the
`[global_defaults]` section anyway.