Correctly handle specified deprecated fields
Previously `dfw::nftables::types::Defaults` was only processed if
`backend_defaults` wasn't unspecified. This had the unintentional
side-effect that the `global_defaults.custom_tables` and
`global_defaults.initilization` fields were not evaluated at all.

To circumvent this, this commit will populate the
`DFW::backend_defaults` field with a default-value in case it hasn't
been specified, which ensures that the core-fields are evaluated even if
`backend_defaults` was not specified by the user.

During this change a further issue was discovered: you were unable to
specify both `backend_defaults` and `initilization` at the same time.
The reason for this is that `initilization` was simply made an alias for
`backend_defaults`, which caused Serde to (rightfully) complain about
the key being specified twice.

To allow that both keys can be specified at the same time, the
`initialization` field was readded as a real field to the `DFW` type,
referencing the `dfw::nftables::types::Initialization` type.

During processing of `dfw::nftables::types::Defaults` the presence of
this field will be handled (and a warning will be shown to the user if
the field was used).

This fixes #265.